v2: full feature parity + functional sidebar UI
Brings the PyQt6 rewrite to feature parity with the legacy app and
wires up a complete, themed sidebar so every operation is reachable
with buttons — while keeping the modern look (dark/light, command
palette, pan/zoom canvas, bottom simulation timeline).

Core (pure, tested):
- trajectory.py: faithful port of the legacy greedy cut-path chaining
  and full-trajectory builder (entry/exit 10mm margins, returns to
  origin, unions, reverse) for G-code parity
- manual_cuts.py: Y/Z/2-point cut lines, segment splitting, snapping,
  contour closing, internal-edge construction
- plates.py: Y×Z grid plate splitting and manual-cut partitioning into
  independent parts (Layer)
- project.py: full state (offset placement, speed, area, reverse,
  manual cuts, batch name) + center-on-bed + machine-coord helper
- dxf_reader.py: $INSUNITS auto-scaling + user scale factor

Services:
- slicer_service: standard trajectory + layer building
- gcode_service: standard and per-plate/part batch export with feed
  from cut speed (mm/s → mm/min)

UI:
- ProjectController: undo/redo-aware operations, signals to refresh
  canvas/sidebar, manual-cut picking, simulation timing, exports
- Sidebar: collapsible sections covering every legacy control —
  load/scale, rotate/auto-height/mirror/fine-rotate/translate/reverse/
  align, speed, plate area + toggles, manual cuts list, plate preview
  navigation, exports
- Canvas scene: bed outline, out-of-bounds highlight, green entry
  marker, dashed manual-cut lines, animated simulation reveal
- Timeline dock: play/pause, scrubber, time/height/dimension read-outs
- main_window: assembles docks, toolbar, menus, palette, toasts,
  QTimer-driven real-time simulation, diagonal-cut canvas picking

Tests: 80 passing (geometry, transforms, clipping, pathing, trajectory,
plates/cuts, gcode, commands, project_io + 4 PyQt6 UI smoke tests).
ruff clean.

+++ appSlicerSegelin/v2/core/pathing.py
@@ -22,7 +22,7 @@ _SNAP_TOL = 1e-6
 
 
 def _key(p: Point, tol: float) -> tuple[int, int]:
-    return (int(round(p.y / tol)), int(round(p.z / tol)))
+    return (round(p.y / tol), round(p.z / tol))
 
 
 def chain_segments(segments: list[Segment], tol: float = _SNAP_TOL) -> list[Path]:
